refactor: solution 도메인 기능별 디렉토리 분리
