+++ backend/src/main/java/com/dajava/backend/domain/solution/dto/SolutionInfoResponse.java
@@ -1,4 +1,4 @@
-package com.dajava.backend.domain.solution;
+package com.dajava.backend.domain.solution.dto;
 
 public record SolutionInfoResponse(
 	String text
