+++ backend/src/main/java/com/dajava/backend/domain/solution/repository/SolutionRepository.java
@@ -1,8 +1,10 @@
-package com.dajava.backend.domain.solution;
+package com.dajava.backend.domain.solution.repository;
 
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
 import com.dajava.backend.domain.register.entity.Register;
+import com.dajava.backend.domain.solution.entity.SolutionEntity;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
